requirements.txt
pyserial~=3.5
bottle~=0.12.18
rrdtool
threading
adafruit-circuitpython-bmp280
